+++ src/plan/generational/immix/gc_work.rs
@@ -1,34 +1,8 @@
 use super::global::GenImmix;
 use crate::plan::generational::gc_work::GenNurseryProcessEdges;
-use crate::plan::TransitiveClosure;
 use crate::policy::gc_work::TraceKind;
-use crate::policy::immix::ImmixSpace;
-use crate::scheduler::GCWorker;
-use crate::util::copy::CopySemantics;
-use crate::util::ObjectReference;
 use crate::vm::VMBinding;
-
-impl<VM: VMBinding> crate::policy::gc_work::UsePolicyProcessEdges<VM> for GenImmix<VM> {
-    type TargetPolicy = ImmixSpace<VM>;
-    const COPY: CopySemantics = CopySemantics::Mature;
-
-    #[inline(always)]
-    fn get_target_space(&self) -> &Self::TargetPolicy {
-        &self.immix
-    }
-
-    #[inline(always)]
-    fn fallback_trace<T: TransitiveClosure>(
-        &self,
-        trace: &mut T,
-        object: ObjectReference,
-        worker: &mut GCWorker<VM>,
-    ) -> ObjectReference {
-        self.gen.trace_object_full_heap::<T>(trace, object, worker)
-    }
-}
-
-use crate::policy::gc_work::PlanProcessEdges;
+use crate::plan::transitive_closure::PlanProcessEdges;
 
 pub struct GenImmixNurseryGCWorkContext<VM: VMBinding>(std::marker::PhantomData<VM>);
 impl<VM: VMBinding> crate::scheduler::GCWorkContext for GenImmixNurseryGCWorkContext<VM> {
