+++ src/policy/largeobjectspace.rs
@@ -118,9 +118,10 @@ impl<VM: VMBinding> Space<VM> for LargeObjectSpace<VM> {
 use crate::scheduler::GCWorker;
 use crate::util::copy::CopySemantics;
 
-impl<VM: VMBinding> crate::plan::transitive_closure::PolicyTraceObject<VM>
+impl<VM: VMBinding> crate::policy::gc_work::PolicyTraceObject<VM>
     for LargeObjectSpace<VM>
 {
+    #[inline(always)]
     fn trace_object<T: TransitiveClosure, const KIND: crate::policy::gc_work::TraceKind>(
         &self,
         trace: &mut T,
