+++ src/policy/lockfreeimmortalspace.rs
@@ -153,9 +153,10 @@ use crate::plan::TransitiveClosure;
 use crate::scheduler::GCWorker;
 use crate::util::copy::CopySemantics;
 
-impl<VM: VMBinding> crate::plan::transitive_closure::PolicyTraceObject<VM>
+impl<VM: VMBinding> crate::policy::gc_work::PolicyTraceObject<VM>
     for LockFreeImmortalSpace<VM>
 {
+    #[inline(always)]
     fn trace_object<T: TransitiveClosure, const KIND: crate::policy::gc_work::TraceKind>(
         &self,
         _trace: &mut T,
